refractored file organization
