+++ cradlepoint/pages/ModalForms/EngagementModalForm.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import Modal from 'react-modal';
-import CPButton from "../components/button/CPButton";
-import styles from '../styles/Modal.module.css'
-import { SmallTextInput, BigTextInput } from "../components/fields/Text";
+import CPButton from "../../components/button/CPButton";
+import styles from '../../styles/Modal.module.css'
+import { SmallTextInput, BigTextInput } from "../../components/fields/Text";
 import { borderLeft } from "@mui/system";
 import PropTypes from 'prop-types';
 import { useRouter } from 'next/router'
