--- cradlepoint/pages/editModalFlow/utils.js
@@ -1,16 +1,0 @@
-const flowType = {
-    ENGAGEMENT: "Engagement",
-    TEST_PLAN: "Test Plan",
-    TEST_CASE: "Test Case",
-    TEST: "Test",
-    RESULT: "Result",
-    DEVICE: "Device",
-}
-
-const modalType = {
-    START: "START",
-    CLONE: "CLONE",
-    SCRATCH: "SCRATCH",
-}
-
-export {flowType, modalType};
